+++ src/AppAny.Quartz.EntityFrameworkCore.Migrations/QuartzModel.cs
@@ -4,7 +4,7 @@ namespace AppAny.Quartz.EntityFrameworkCore.Migrations
 {
   public class QuartzModel
   {
-    public string? Prefix { get; set; } = "qrtz_";
+    public string? Prefix { get; set; } = "QRTZ_";
     public string? Schema { get; set; }
     public Action<EntityTypeConfigurationContext>? EntityTypeConfigurations { get; set; }
   }
